// src/generated/models/session-response-data.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Finatic FastAPI Backend
 * FinaticAPI REST API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SessionStatus } from './session-status';

/**
 * Response data for session operations.
 */
export interface SessionResponseData {
    /**
     * Session ID
     */
    'session_id': string;
    /**
     * Company ID
     */
    'company_id': string;
    /**
     * Session status
     */
    'status': SessionStatus;
    /**
     * Session expiration time
     */
    'expires_at': string;
    'user_id'?: string | null;
}